cpuidle: teo: Add a timer to avoid staying in WFI for too long on arm64 It is often the case that a CPU may remain in a shallow idle state longer than the cpuidle governor originally predicted, leading to lost energy savings had the CPU been placed into a deeper idle state.
Since the tick is never stopped for WFI in the TEO governor, there is at
least an upper limit to how long a CPU may remain in WFI. However,
remaining in WFI for an entire tick is suboptimal compared to arming a
timer to exit WFI earlier.

That being said, it is difficult to implement such a timeout in a generic
way, since it is innately tied to CPU architectural details.

Therefore, implement an idle timeout in arm64 using TEO in order to improve
CPU idle energy savings significantly. A timeout of 1 ms is empirically
found to provide the best energy saving benefit, and should be applicable
to most arm64 CPUs since WFI exit latency is generally quite low.

On Tensor G3, this consistently reduces CPU + RAM power consumption by at
least 30 mW in a light gaming load scenario (Temple Run 2).

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>